Fix null issue in code-sharer

+++ microservices/code-sharer/controllers/code-sockets/code-session-socketio.js
@@ -34,17 +34,20 @@ const removeUser = async (socket) => {
     const room_id = socketSession.room_id;
 
     let codeSession = await codeSessionDao.findCodeSession(room_id);
-    codeSession.num_active_users -= 1;
 
-    if (codeSession.num_active_users === 0) {
-      await codeSessionDao.deleteCodeSession(room_id);
-      // Call the API to push code session to MongoDB and S3
-    } else {
-      await codeSessionDao.updateCodeSession(room_id, codeSession);
+    if (codeSession) {
+      codeSession.num_active_users -= 1;
+
+      if (codeSession.num_active_users === 0) {
+        await codeSessionDao.deleteCodeSession(room_id);
+        // Call the API to push code session to MongoDB and S3
+      } else {
+        await codeSessionDao.updateCodeSession(room_id, codeSession);
+      }
+
+      await socketSessionDao.deleteSocketSession(socket.id);
+      socket.leave(room_id);
     }
-
-    await socketSessionDao.deleteSocketSession(socket.id);
-    socket.leave(room_id);
   }
 };
